allcator: change &mut self -> &mut self to mut self -> self
Previous constraint implemented &mut self -> &mut self for changing max,
min, etc.

But this cause problem for code like
```
let constraint = Constraint::new(...).max(...).min(...);
```

This code creates a Constraint and returns the mut reference to it, but
there will be no one take the ownership of that Constraint so it'll be
dropped and cannot be used later. Since temporary value is freed at the end
of this statement, it is not a ideal case for us.

Signed-off-by: Chao Wu <chaowu@linux.alibaba.com>