db-boot: arm add layout.rs
Align the x86 architecture, put arm layout.rs in the db-boot directory,
and db-boot depends on db-arch.

Since the gic module is stored in the db-arch directory, and the
virtualized gic needs to specify the mmio address space of the gic.
Therefore, the mmio address space of gic is first defined in the gic
module, and on this basis, other contents in layout.rs are continued to
be constructed.

Signed-off-by: jingshan.lx <jingshan@linux.alibaba.com>