dbs-boot: arm add layout.rs
Align the x86 architecture, put arm layout.rs in the dbs-boot directory,
and dbs-boot depends on dbs-arch.

Since the gic module is stored in the dbs-arch directory, and the
virtualized gic needs to specify the mmio address space of the gic.
Therefore, the mmio address space of gic is first defined in the gic
module, and on this basis, other contents in layout.rs are continued to
be constructed.

In addition, we define IRQ_BASE and IRQ_MAX also depend on the relevant
definitions in gic.

Signed-off-by: jingshan.lx <jingshan@linux.alibaba.com>