virtio-fs: add VirtioRegionHandler to support DAX memory.
When a guest application writes DAX memory to a no-dax virtiofs file,
for example:
```
    void *dax_ptr = (void*)mmap(NUMM, 4096, PORT, MAP_SHARED, dax_fd, 0);
    write(no_dax_fd, dax_ptr, 4096);
```

dragonball will coredump.

The reason for coredump is that Virtiofs does not insert the DAX region into vm_as.
Thus, vm_as cannot find the region corresponding to the DAX address in
Reader::new(mem, desc.clone()).

This patch inserts DAX Region into vm_as via callback in VirtioFs::set_resource to
resolve the coredump.

DAX Region also inserted into address_space so that the e820 table can be set correctly.
The e820 table is a table that describes guest memory prepared before the guest startup.
we need to config the correct guest memory address and length in the table. The virtio
device memory belongs to the MMIO space and does not belong to the Guest Memory space.
Therefore, it cannot be configured into the e820 table. When creating AddressSpaceRegion we
use AddressSpaceRegionType::ReservedMemory type, in this way, address_space will know that
this region a special memory, it will don't put the this memory in e820 table.

Signed-off-by: gexuyang <gexuyang@linux.alibaba.com>