dbs-upcall: introduce upcall
dbs-upcall is a direct communication tool between VMM and guest
developed upon vsock. The server side of the upcall is a driver in guest
kernel (kernel patches are needed for this feature) and it'll start to
serve the requests after the kernel starts. And the client side is in
VMM, it is a thread that communicates with vsock through uds.

We have accomplished device hotplug / hot-unplug directly through upcall
in order to avoid virtualization of ACPI to minimize virtual machines'
overhead. And there could be many other usage through this direct
communication channel.

Signed-off-by: Chao Wu <chaowu@linux.alibaba.com>